FEATURE: Support Gitlab Blobs; fix specs.
