+++ lib/onebox/engine.rb
@@ -183,3 +183,4 @@ require_relative "engine/wechat_mp_onebox"
 require_relative "engine/cloudapp_onebox"
 require_relative "engine/wistia_onebox"
 require_relative "engine/instagram_onebox"
+require_relative "engine/gitlab_blob_onebox"
