+++ spec/lib/onebox/engine/github_blob_onebox_spec.rb
@@ -2,8 +2,8 @@ require "spec_helper"
 
 describe Onebox::Engine::GithubBlobOnebox do
   before(:all) do
-    @link = "https://github.com/discourse/discourse/blob/master/lib/oneboxer/github_blob_onebox.rb"
-    fake("https://raw.github.com/discourse/discourse/master/lib/oneboxer/github_blob_onebox.rb", response(described_class.onebox_name))
+    @link = "https://github.com/discourse/onebox/blob/master/lib/onebox/engine/github_blob_onebox.rb"
+    fake("https://raw.github.com/discourse/onebox/master/lib/onebox/engine/github_blob_onebox.rb", response(described_class.onebox_name))
   end
 
   include_context "engines"
